Remove test superclass
